GP2-1398_Where-to-export_split_data_table_loading
