+++ core/templates/core/base.html
@@ -53,6 +53,7 @@
           apiFundingCreditOptionsDeleteUrl: '{% url "exportplan:api-funding-credit-options-delete" %}',
           countriesBySectorsDataUrl: '{% url "exportplan:ajax-recommended-countries-data" %}',
           apiComTradeDataUrl: '{% url "core:api-comtrade-data" %}',
+          apiCountryDataUrl: '{% url "core:api-country-data" %}',
           removeSectorUrl: '{% url "exportplan:api-remove-sector" %}',
           countryAgeGroupDataUrl: '{% url "exportplan:api-target-age-country-population-data" %}',
           populationByCountryUrl: '{% url "exportplan:api-population-data-by-country" %}',
