+++ core/urls.py
@@ -69,4 +69,5 @@ urlpatterns = [
     path('api/create-token/', skip_ga360(views_api.CreateTokenView.as_view()), name='api-create-token'),
     path('api/check/', skip_ga360(views_api.CheckView.as_view()), name='api-check'),
     path('api/data-service/comtrade/', skip_ga360(views_api.ComTradeDataView.as_view()), name='api-comtrade-data'),
+    path('api/data-service/countrydata/', skip_ga360(views_api.CountryDataView.as_view()), name='api-country-data'),
 ]
